+++ src/hardware_layer/sensor_access/include/sensor_access/mock_pos_sensor.h
@@ -7,9 +7,9 @@
 class MockPosSensor : public PosSensorInterface
 {
 public:
-  MOCK_METHOD0(getX, float());
-  MOCK_METHOD0(getY, float());
-  MOCK_METHOD0(getTheta, float());
+  MOCK_METHOD0(getX, double());
+  MOCK_METHOD0(getY, double());
+  MOCK_METHOD0(getTheta, double());
   MOCK_METHOD0(receiveData, ReadableSensors::ReadStatus());  // this is an "uninteresting function"
 };
