+++ src/localization/include/super_localizer/super_localizer_helper.h
@@ -4,6 +4,6 @@ namespace LocalizerInterface
 LocalizerInterface::stateVector diff(LocalizerInterface::stateVector lhs, LocalizerInterface::stateVector rhs);
 LocalizerInterface::stateVector multiply(LocalizerInterface::stateVector lhs, LocalizerInterface::stateVector rhs);
 LocalizerInterface::stateVector addFromModel(LocalizerInterface::stateVector lhs, LocalizerInterface::stateVector rhs,
-                                             float dt, bool imu);
-LocalizerInterface::stateVector initState(float xi, float yi, float theta);
+                                             double dt, bool imu);
+LocalizerInterface::stateVector initState(double xi, double yi, double theta);
 }
